feat
The function initializes a variable numAccounts to keep track of the number of created accounts and dynamically allocates memory for an array of Account pointers called accounts.

A while loop is used to repeatedly display the main menu and prompt the user for their choice.

The user can choose to create an account, log in, or exit the program. Based on the user's choice, the relevant functionality is executed.

For account creation (case '1'), the user is prompted to enter an account number and an account holder name. The createAccount() function is called to create a new account with the provided details.

If the account creation is successful, the new account is added to the accounts array using dynamic memory allocation to resize the array. The numAccounts variable is incremented accordingly.

If the maximum number of accounts is reached, an error message is displayed, indicating that the user cannot create more accounts.